Variation update

+++ app/Http/Controllers/Admin/ProductController.php
@@ -192,16 +192,19 @@ class ProductController extends Controller
 
         foreach ($request->get('variations') as $id => $variation)
         {
-            $variationObj = ProductVariation::find($id);
-            $variationObj->title = $variation['title'] ?? '';
-            $variationObj->fixed_price = $variation['fixed_price'] ?? 0;
-            $variationObj->adding_price = $variation['adding_price'] ?? 0;
-            $variationObj->slug = Str::slug($variation['title']);
-            $variationObj->save();
-
-            if (empty($variation['title']))
+            if (isset($variation['title']))
             {
-                $variationObj->delete();
+                $variationObj = ProductVariation::find($id);
+                $variationObj->title = $variation['title'] ?? '';
+                $variationObj->fixed_price = $variation['fixed_price'] ?? 0;
+                $variationObj->adding_price = $variation['adding_price'] ?? 0;
+                $variationObj->slug = Str::slug($variation['title']);
+                $variationObj->save();
+
+                if (empty($variation['title']))
+                {
+                    $variationObj->delete();
+                }
             }
         }
